Change Gamestate to dds file
