+++ bridge/double_dummy_solver.py
@@ -1,4 +1,43 @@
-from .models import GameState, Player
+from .models import Player, Trick
+
+
+class GameState:
+    def __init__(self, board, trump, starting_player):
+        self.board = board
+        self.trump = trump
+        self.next_player = starting_player
+        self.previous_tricks = []
+        self.current_trick = Trick(trump, starting_player)
+        self.ns_trick_count = 0
+
+    def candidate_cards(self):
+        current_suit = self.current_trick.trick_suit()
+        return self.board.get_hand(self.next_player).candidate_cards(current_suit)
+
+    def play_card(self, card):
+        self.board.get_hand(self.next_player).play_card(card)
+
+        trick_winner = self.current_trick.play_card(card)
+        if trick_winner is None:
+            self.next_player = self.next_player.next_player()
+        else:
+            self.next_player = trick_winner
+            self.previous_tricks.append(self.current_trick)
+            self.current_trick = Trick(self.trump, trick_winner)
+            if trick_winner == Player.NORTH or trick_winner == Player.SOUTH:
+                self.ns_trick_count += 1
+
+    def step_back(self):
+        last_card = self.current_trick.step_back()
+        if last_card is None:
+            self.current_trick = self.previous_tricks.pop()
+            if self.next_player == Player.NORTH or self.next_player == Player.SOUTH:
+                self.ns_trick_count -= 1
+            self.next_player = self.current_trick.starting_player.previous_player()
+            last_card = self.current_trick.step_back()
+        else:
+            self.next_player = self.next_player.previous_player()
+        self.board.get_hand(self.next_player).add_card(last_card)
 
 
 def double_dummy_solver(board, trump, starting_player):
@@ -12,7 +51,7 @@ def double_dummy_solver_using_gameState(gameState):
     if len(candidate_cards) == 0:
         return gameState.ns_trick_count
 
-    if next_player == Player.NORTH or next_player == Player.SOUTH:
+    if next_player in [Player.NORTH, Player.SOUTH]:
         max_trick = 0
         for card in candidate_cards:
             gameState.play_card(card)
