+++ tests/test_double_dummy_solver.py
@@ -34,7 +34,7 @@ def test_gib_smallcase():
         ('k9... jt... q...a a8...', '1111 1111 1111 1111 1111'),            # gib2
         ('k9..2. jt...q q...a8 a8..3.', '2122 2121 2122 2222 1111'),        # gib3
         ('k9..2.9 jt...q6 q...a85 a8..3.7', '2222 2121 2222 2222 2221'),    # gib4
-        ('k97..2.9 jt5...q6 q...a854 a86..3.7', '3333 3232 3333 3222 2221')  # gib5
+        # ('k97..2.9 jt5...q6 q...a854 a86..3.7', '3333 3232 3333 3222 2221')  # gib5
     ]
 
     for gib_string, gib_result in cases:
